Implemented star based rating in review; Removed unecessary check for user in controller, takes directly from the View; Changed from int to double to take half grading;

+++ Cocktail_Magician_Services/DTO/CocktailRatingDTO.cs
@@ -8,6 +8,6 @@ namespace Cocktail_Magician_Services.DTO
     {
         public string UserId { get; set; }
         public string CocktailId { get; set; }
-        public int Grade { get; set; }
+        public double Grade { get; set; }
     }
 }
